nfc: ndef: enable encoding of long uri messages
The NDEF URI Message module now supports long URI payload encoding.
It is possible to encode URIs longer than 255 bytes.

+++ include/nfc/ndef/uri_rec.h
@@ -80,7 +80,7 @@ struct nfc_ndef_uri_rec_payload {
 	/** Pointer to a URI string. */
 	uint8_t const *uri_data;
 	/** Length of the URI string. */
-	uint8_t uri_data_len;
+	uint16_t uri_data_len;
 };
 
 /**
